+++ src/main/java/edu/clemson/cs/r2jt/init2/file/ConcreteFile.java
@@ -10,7 +10,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.r2jt.init2.model;
+package edu.clemson.cs.r2jt.init2.file;
 
 /**
  * TODO: Description for this class
